+++ device-minion-runtime/conanfile.py
@@ -1,6 +1,6 @@
 from conan import ConanFile
 from conan.errors import ConanInvalidConfiguration
-from conan.tools.cmake import CMake, CMakeToolchain, CMakeDeps
+from conan.tools.cmake import CMake, CMakeToolchain, CMakeDeps, cmake_layout
 from conan.tools.files import rmdir
 import os
 import textwrap
@@ -10,7 +10,8 @@ required_conan_version = ">=1.53.0"
 
 class DeviceMinionRuntimeConan(ConanFile):
     name = "device-minion-runtime"
-    url = "https://gitlab.esperanto.ai/software/device-minion-runtime"
+    url = "git@gitlab.com:esperantotech/software/device-minion-runtime.git"
+    homepage = "https://gitlab.com/esperantotech/software/device-minion-runtime"
     description = "minion-rt runtime"
     license = "Esperanto Technologies"
 
@@ -22,17 +23,18 @@ class DeviceMinionRuntimeConan(ConanFile):
         "warnings_as_errors": True
     }
 
-    scm = {
-        "type": "git",
-        "url": "git@gitlab.esperanto.ai:software/device-minion-runtime.git",
-        "revision": "auto",
-    }
-
     python_requires = "conan-common/[>=1.1.0 <2.0.0]"
 
     def set_version(self):
         get_version = self.python_requires["conan-common"].module.get_version
         self.version = get_version(self, "deviceMinionRuntime")
+
+    def export(self):
+        register_scm_coordinates = self.python_requires["conan-common"].module.register_scm_coordinates
+        register_scm_coordinates(self)
+
+    def layout(self):
+        cmake_layout(self)
 
     def configure(self):
         self.settings.rm_safe("compiler.libcxx")
@@ -52,6 +54,14 @@ class DeviceMinionRuntimeConan(ConanFile):
 
     def build_requirements(self):
         self.tool_requires("cmake-modules/[>=0.4.1 <1.0.0]")
+
+    def export_sources(self):
+        copy_sources_if_scm_dirty = self.python_requires["conan-common"].module.copy_sources_if_scm_dirty
+        copy_sources_if_scm_dirty(self)
+
+    def source(self):
+        get_sources_if_scm_pristine = self.python_requires["conan-common"].module.get_sources_if_scm_pristine
+        get_sources_if_scm_pristine(self)
 
     def validate(self):
         if self.settings.arch != "rv64":